feat: Display the new numbers upon clicking button
Added Event Listeners to each button to either update the display or
reassign the number in the display based on if the number is
an answer/the number is a 0 or alert a message when trying to break
the number limit.

Also upon each button click, we store the value in a variable so we
can give it to a number variable for operation.