Add a dot to allow users to type in decimals
We need to check if the number is sensitive(is an answer or num1 after
typing in an operator) because we can't add the dot to a number that's
supposed to be immutable, and add the dot to a 0 if so.

Else just check if there's already a dot and add or don't based
on the answer.