Add Keyboard Support
Added a listener for key down strokes(not key up because enter caused
a lot of problems - pressed a button and acted as "=" at the same time
because of the previous mouse focus on any button) on window and not
document object.

Then execute the updated functions(which now include keyboard support)
and get the same result as when pressing buttons.